[feat]: 많은 것을 수정 - 기존에는 db에 회원카트, 비회원카트 모두 저장했는데 비회원카트는 세션아이디를 받아서 레디스에 저장

+++ src/main/java/com/dooray/bookstorecarts/entity/CartItem.java
@@ -14,11 +14,11 @@ public class CartItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cart_item_Id")
-    private int id;
+    private Long id;
     @Column(name = "book_id")
-    private int bookId;
+    private Long bookId;
     @Column(name = "quantity")
-    private int quantity;
+    private Long quantity;
     @ManyToOne
     @JoinColumn(name = "cart_id")
     private Cart cart;
